requirements.txt
aiofiles
anthropic
boto3
python-dotenv
playsound
pydantic
pygame
PyObjC
